+++ packages/kiarina-agi-tool/src/kiarina/agi/post_hook/_types/post_hook_output.py
@@ -3,4 +3,4 @@ from typing import TypeAlias
 
 from kiarina.agi.event import Event
 
-PostHookOutput: TypeAlias = None | Awaitable[None] | AsyncIterator[Event]
+PostHookOutput: TypeAlias = Awaitable[None] | AsyncIterator[Event] | None