+++ packages/kiarina-lib-redisearch/src/kiarina/lib/redisearch/_core/operations/reset_index.py
@@ -27,7 +27,7 @@ def reset_index(
 def reset_index(
     mode: Literal["sync", "async"],
     ctx: RedisearchContext,
-) -> None | Awaitable[None]:
+) -> Awaitable[None] | None:
 
     def _log_delete_index() -> None:
         logger.info("Deleting existing index '%s'", ctx.settings.index_name)