+++ packages/kiarina-utils-file/src/kiarina/utils/file/_core/utils/write_json_list.py
@@ -38,7 +38,7 @@ def write_json_list(
     indent: int = 2,
     ensure_ascii: bool = False,
     sort_keys: bool = False,
-) -> None | Awaitable[None]:
+) -> Awaitable[None] | None:
     json_text = json.dumps(
         json_list, indent=indent, ensure_ascii=ensure_ascii, sort_keys=sort_keys
     )