+++ packages/kiarina-utils-file/src/kiarina/utils/file/_core/utils/write_yaml_list.py
@@ -36,7 +36,7 @@ def write_yaml_list(
     *,
     allow_unicode: bool = True,
     sort_keys: bool = False,
-) -> None | Awaitable[None]:
+) -> Awaitable[None] | None:
     yaml_text = yaml.dump(yaml_list, allow_unicode=allow_unicode, sort_keys=sort_keys)
 
     def _sync() -> None: